added a simple  'text color changing with click' project .
This folder contains two files index.html and t.js. The script file contains a simple immediatly invoked function expression which also contains an inner function attached to an event listener . On clicking anywhere on the screen function is called which toggles the text color from blue to red and vice versa.
Its primary pupose was to understand the concept of closures in javascript.